moved metadata exception into exceptions/
